+++ lib/sample_uploader/utils/importer.py
@@ -1,7 +1,8 @@
 import pandas as pd
 import warnings
 import os
-from installed_clients.WorkspaceClient import Workspace
+import csv
+
 from sample_uploader.utils.sample_utils import (
     get_sample,
     save_sample,
@@ -14,7 +15,7 @@ from sample_uploader.utils.sample_utils import (
 )
 from sample_uploader.utils.transformations import FieldTransformer
 from sample_uploader.utils.parsing_utils import upload_key_format
-from sample_uploader.utils.mappings import SAMP_SERV_CONFIG
+from sample_uploader.utils.mappings import SAMP_SERV_CONFIG, CORE_FIELDS, NON_PREFIX_TO_PREFIX
 from sample_uploader.utils.misc_utils import get_workspace_user_perms
 from sample_uploader.utils.samples_content_warning import SampleContentWarning, SampleContentWarningContext
 
@@ -22,6 +23,47 @@ from sample_uploader.utils.samples_content_warning import SampleContentWarning, 
 REQUIRED_COLS = {'name'}
 REGULATED_COLS = ['name', 'id', 'parent_id']
 NOOP_VALS = ['ND', 'nd', 'NA', 'na', 'None', 'n/a', 'N/A', 'Na', 'N/a', '-']
+
+
+def find_header_row(sample_file, file_format):
+    if not os.path.isfile(sample_file):
+        # try prepending '/staging/' to file and check then
+        if os.path.isfile(os.path.join('/staging', sample_file)):
+            sample_file = os.path.join('/staging', sample_file)
+        else:
+            raise ValueError(f"input file {sample_file} does not exist.")
+
+    header_row_index = 0
+    if file_format.lower() == "sesar":
+
+        if sample_file.endswith('.tsv') or sample_file.endswith('.csv'):
+            reader = pd.read_csv(sample_file, sep=None, iterator=True)
+            inferred_sep = reader._engine.data.dialect.delimiter
+            with open(sample_file) as f:
+                first_line = f.readline()
+                second_line = f.readline()
+
+            # when an extra header presents,
+            # the first line (SESAR header line) should have an unequal number of columns than the second line (the real header line).
+            # TODO: this function will fail if the extra header line happens to have exactly the same number of columns as the real header line
+            non_empty_header = [i for i in first_line.split(inferred_sep) if i not in ['', '\n']]
+            if len(non_empty_header) != len(second_line.split(inferred_sep)):
+                header_row_index = 1
+
+        elif sample_file.endswith('.xls') or sample_file.endswith('.xlsx'):
+            df = pd.read_excel(sample_file)
+
+            # when an extra header presents,
+            # all of the unmatched columns are indexed as 'Unnamed xx'.
+            unnamed_cols = [i for i in df.columns if 'Unnamed' in i]
+            if len(unnamed_cols) > 0:
+                header_row_index = 1
+        else:
+            raise ValueError(f"File {os.path.basename(sample_file)} is not in "
+                             f"an accepted file format, accepted file formats "
+                             f"are '.xls' '.csv' '.tsv' or '.xlsx'")
+
+    return header_row_index
 
 
 def validate_params(params):
@@ -37,7 +79,7 @@ def validate_params(params):
         else:
             raise ValueError(f"input file {sample_file} does not exist.")
     if params.get('name_field'):
-        try: 
+        try:
             upload_key_format(params.get('name_field'))
         except SampleContentWarning as e:
             raise ValueError("Invalid ID field in params: {e.message}")
@@ -86,7 +128,8 @@ def _produce_samples(
     if not REQUIRED_COLS.issubset(df.columns):
         raise ValueError(
             f'Required "name" column missing from input. Use "name" or '
-            f'an alias (sample name", "sample id", "samplename", "sampleid")'
+            f'an alias ("sample name", "sample id", "samplename", "sampleid") '
+            f'Existing fields ({df.columns})'
         )
 
     def _get_existing_sample(name, kbase_sample_id):
@@ -241,9 +284,11 @@ def _save_samples(samples, acls, sample_url, token):
     return saved_samples
 
 
-def format_input_file(df, columns_to_input_names, aliases):
+def format_input_file(df, params, columns_to_input_names, aliases):
     # change columns to upload format
     columns_to_input_names = {}
+
+    # set column names into 'upload_key_format'.
     for col_idx, col_name in enumerate(df.columns):
         renamed = upload_key_format(col_name)
         if renamed not in columns_to_input_names:
@@ -259,6 +304,21 @@ def format_input_file(df, columns_to_input_names, aliases):
 
     df = df.rename(columns={columns_to_input_names[col]: col for col in columns_to_input_names})
     df.replace({n: None for n in NOOP_VALS}, inplace=True)
+
+    # TODO: Make sure to check all possible name fields, even when not parameterized
+    if params.get('name_field'):
+        name_field = upload_key_format(params.get('name_field'))
+        if name_field not in list(df.columns):
+            raise ValueError(
+                f"The expected name field column \"{name_field}\" could not be found. "
+                "Adjust your parameters or input such that the following are correct:\n"
+                f"- File Format: {params.get('file_format')} (the format to which your sample data conforms)\n"
+                f"- ID Field: {params.get('name_field','name')}\n (the header of the column containing your names)\n"
+                f"- Headers Row: {params.get('header_row_index')} (the row # where column headers are located in your spreadsheet)"
+            )
+        # here we rename whatever the id field was/is to "id"
+        columns_to_input_names["name"] = columns_to_input_names.pop(name_field)
+        df.rename(columns={name_field: "name"}, inplace=True)
 
     map_aliases = {}
     for key, key_aliases in aliases.items():
@@ -278,6 +338,32 @@ def format_input_file(df, columns_to_input_names, aliases):
 
     if map_aliases:
         df = df.rename(columns=map_aliases)
+
+    file_format = params.get('file_format').lower()
+
+    prefix_map = {}
+    for col in df.columns:
+        if ':' in col:
+            continue
+        fields = NON_PREFIX_TO_PREFIX.get(col, [])
+        target_field = None
+        for field in fields:
+            # choose the format that fits if it exists.
+            if field.split(":")[0] == file_format:
+                target_field = field
+                break
+            else:
+                if col in CORE_FIELDS:
+                    target_field = col
+                else:
+                    target_field = field
+        if target_field:
+            prefix_map[col] = target_field
+            if col in columns_to_input_names:
+                val = columns_to_input_names.pop(col)
+                columns_to_input_names[target_field] = val
+    if prefix_map:
+        df = df.rename(columns=prefix_map)
 
     return df, columns_to_input_names
 
@@ -305,30 +391,20 @@ def import_samples_from_file(
         sample_file = validate_params(params)
         ws_name = params.get('workspace_name')
         df = load_file(sample_file, header_row_index, date_columns)
-        df, columns_to_input_names = format_input_file(df, {}, aliases)
 
-        # TODO: Make sure to check all possible name fields, even when not parameterized
-        if params.get('name_field'):
-            name_field = upload_key_format(params.get('name_field'))
-            if name_field not in list(df.columns):
-                raise ValueError(
-                    f"The expected name field column \"{name_field}\" could not be found. "
-                    "Adjust your parameters or input such that the following are correct:\n"
-                    f"- File Format: {params.get('file_format')} (the format to which your sample data conforms)\n"
-                    f"- ID Field: {params.get('name_field','name')}\n (the header of the column containing your names)\n"
-                    f"- Headers Row: {params.get('header_row_index')} (the row # where column headers are located in your spreadsheet)"
-                )
-            # here we rename whatever the id field was/is to "id"
-            columns_to_input_names["name"] = columns_to_input_names.pop(name_field)
-            df.rename(columns={name_field: "name"}, inplace=True)
+        file_format = params['file_format'].lower()
+        if 'sample_template' not in df:
+            df['sample_template'] = file_format.upper()
+
+        df, columns_to_input_names = format_input_file(df, params, {}, aliases)
 
         if not errors.get(severity='error'):
-            if params['file_format'].lower() in ['sesar', "enigma"]:
+            if file_format in ['sesar', "enigma"]:
                 if 'material' in df.columns:
-                    df.rename(columns={"material": params['file_format'].lower() + ":material"}, inplace=True)
+                    df.rename(columns={"material": file_format + ":material"}, inplace=True)
                     val = columns_to_input_names.pop("material")
-                    columns_to_input_names[params['file_format'].lower() + ":material"] = val
-            if params['file_format'].lower() == "kbase":
+                    columns_to_input_names[file_format + ":material"] = val
+            if file_format == "kbase":
                 if 'material' in df.columns:
                     df.rename(columns={"material": "sesar:material"}, inplace=True)
                     val = columns_to_input_names.pop("material")