+++ test/core_tests.py
@@ -16,6 +16,7 @@ from sample_uploader.authclient import KBaseAuth as _KBaseAuth
 from sample_uploader.utils.sample_utils import get_sample
 from installed_clients.WorkspaceClient import Workspace
 
+
 class sample_uploaderTest(unittest.TestCase):
 
     @classmethod
@@ -180,7 +181,7 @@ class sample_uploaderTest(unittest.TestCase):
         samples = [get_sample(sample_info, self.sample_url, self.ctx['token'])
                    for sample_info in samples_info]
 
-        sample_igsns = [sample['node_tree'][0]['meta_controlled']['igsn']['value'] for sample in samples]
+        sample_igsns = [sample['node_tree'][0]['meta_controlled']['sesar:igsn']['value'] for sample in samples]
         assert set(sample_igsns) == set(igsns)
 
         expected_sample_names = ['PB-Low-5', 'ww163e', 'Core 1-1*-1M']
@@ -203,7 +204,7 @@ class sample_uploaderTest(unittest.TestCase):
         samples = [get_sample(sample_info, self.sample_url, self.ctx['token'])
                    for sample_info in samples_info]
 
-        sample_igsns = [sample['node_tree'][0]['meta_controlled']['igsn']['value'] for sample in samples]
+        sample_igsns = [sample['node_tree'][0]['meta_controlled']['sesar:igsn']['value'] for sample in samples]
         assert set(sample_igsns) == set(igsns)
 
         expected_sample_names = ['PB-Low-5', 'ww163e', 'Core 1-1*-1M']
@@ -226,7 +227,7 @@ class sample_uploaderTest(unittest.TestCase):
 
         sample = get_sample(samples_info[0], self.sample_url, self.ctx['token'])
 
-        sample_igsn = sample['node_tree'][0]['meta_controlled']['igsn']['value']
+        sample_igsn = sample['node_tree'][0]['meta_controlled']['sesar:igsn']['value']
         assert sample_igsn == igsns
 
         expected_sample_names = 'PB-Low-5'